setup CoreData made

+++ ToDoList/ModelCacheData/ToDoItem.swift
@@ -87,10 +87,10 @@ extension ToDoItem {
     
     var likeElement: Element {
         
-        var deadlineValue: Int?
+        var deadlineValue: Int64?
         
         if let deadline = deadline {
-            deadlineValue = Int(deadline.timeIntervalSince1970)
+            deadlineValue = Int64(deadline.timeIntervalSince1970)
         }
         
         return Element(id: id, text: text, deadline: deadlineValue, importance: importance.rawValue, done: isDone, color: nil, createdAt: Int(dateCreated.timeIntervalSince1970), changedAt: Int(dateChanged?.timeIntervalSince1970 ?? Date().timeIntervalSince1970), lastUpdatedBy: UIDevice.current.identifierForVendor!.uuidString)
